Modified a lot of things: read description
Reimplemented the Player class (TO BE COMPLETED). Corrected a bug in the insert method of the Shelf class. Added the method getRandomType in the ItemType enum for testing reasons. Reimplemented the ShelfTest class because were obsolete for the new implementation. Refectored the Game class to reflect the Player class implementation. Added GameTest class