fixed small, but important bug: now you can play again
in Game class, the numberOfPlayers attribute must be set to -1. Setting it to 0 throws IllegalStateException when trying to call the addPlayer method, because the players ArrayList is empty when initializing the class, so the condition numbersOfPlayers == players.size() is met,. Setting it to 1, allows a single player to play alone.